backend/pyproject.toml
[tool.poetry]
name = "pet-project"
version = "0.1.0"
description = ""
authors = ["Yurochka <yurii.yurochka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic-settings = "^2.1.0"
motor = "^3.3.2"
motor-types = "^1.0.0b4"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.14"
mypy = "^1.8.0"
pre-commit = "^3.6.0"

[tool.ruff]
line-length=79
target-version = 'py312'

[tool.ruff.lint]
select = [
  # Pyflakes
  "F",
  # pycodestyle
  "E4",
  "E7",
  "E9",
  # isort
  "I",
  # pep8-naming
  "N",
]

[tool.mypy]
python_version = "3.12"
strict = true
mypy_path = "./src/"
allow_untyped_globals = true
check_untyped_defs = true
disallow_any_decorated = true
disallow_any_generics = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
ignore_missing_imports = false
show_error_codes = true
warn_return_any = true
warn_unused_configs = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
